refactor: move area chart example to typescript
