+++ package.json
@@ -42,6 +42,7 @@
     "html-webpack-plugin": "^4.3.0",
     "storybook-addon-jsx": "^7.3.14",
     "style-loader": "^1.2.1",
+    "typescript": "^4.8.4",
     "webpack": "^4.44.1",
     "webpack-cli": "^3.3.12",
     "webpack-dev-server": "^3.11.0"
