+++ src/types/data/DataProps.ts
@@ -1,0 +1,5 @@
+
+
+export type TimeSeriesData = { date: string; value: number; }
+
+export type CategoricalData = { category: string; value: number; }
